MinecartTerm-System in /core, And-Links
We moved the whole MinecartTerm-System to the /core, so you don't need
to hook into the BasicExpressions-plugin for adding custom terms
anymore. We also fixed some bugs in the users of the MinecartTerms
(Intersection and Sensor) and added a system for linking terms to Ands
(so every of the conditions has to be true) with the "&"-character.